integrate Q methodology, concern extraction
I am trying to integrate Q methodology into the chatbot, having a bit of trouble but this version runs. There are some bugs, we need to make it so it asks one concern at a time. If we can make the UI similar to Borys and the UX devs that would be best.